Make a directory for the dynamic_types
