+++ ddsrecorder_participants/include/ddsrecorder_participants/common/types/dynamic_types/v2/DynamicTypesCollectionPubSubTypes.hpp
@@ -29,7 +29,7 @@
 #include <fastdds/rtps/common/SerializedPayload.h>
 #include <fastrtps/utils/md5.h>
 
-#include <ddsrecorder_participants/common/types/v2/DynamicTypesCollection.hpp>
+#include <ddsrecorder_participants/common/types/dynamic_types/v2/DynamicTypesCollection.hpp>
 
 
 #if !defined(GEN_API_VER) || (GEN_API_VER != 2)
