+++ ddsrecorder_participants/include/ddsrecorder_participants/recorder/mcap/McapHandler.hpp
@@ -40,9 +40,9 @@
 #include <ddsrecorder_participants/recorder/mcap/McapHandlerConfiguration.hpp>
 
 #if FASTRTPS_VERSION_MAJOR <= 2 && FASTRTPS_VERSION_MINOR < 13
-    #include <ddsrecorder_participants/common/types/v1/DynamicTypesCollection.hpp>
+    #include <ddsrecorder_participants/common/types/dynamic_types/v1/DynamicTypesCollection.hpp>
 #else
-    #include <ddsrecorder_participants/common/types/v2/DynamicTypesCollection.hpp>
+    #include <ddsrecorder_participants/common/types/dynamic_types/v2/DynamicTypesCollection.hpp>
 #endif // if FASTRTPS_VERSION_MAJOR <= 2 && FASTRTPS_VERSION_MINOR < 13
 
 namespace eprosima {
