+++ ddsrecorder_participants/src/cpp/common/types/dynamic_types/v2/DynamicTypesCollection.idl
