+++ ddsrecorder_participants/src/cpp/recorder/mcap/McapHandler.cpp
@@ -41,12 +41,12 @@
     #include <fastcdr/Cdr.h>
     #include <fastcdr/FastBuffer.h>
     #include <fastcdr/FastCdr.h>
-    #include <ddsrecorder_participants/common/types/v1/DynamicTypesCollection.hpp>
-    #include <ddsrecorder_participants/common/types/v1/DynamicTypesCollectionPubSubTypes.hpp>
+    #include <ddsrecorder_participants/common/types/dynamic_types/v1/DynamicTypesCollection.hpp>
+    #include <ddsrecorder_participants/common/types/dynamic_types/v1/DynamicTypesCollectionPubSubTypes.hpp>
 #else
     #include <fastdds/rtps/common/CdrSerialization.hpp>
-    #include <ddsrecorder_participants/common/types/v2/DynamicTypesCollection.hpp>
-    #include <ddsrecorder_participants/common/types/v2/DynamicTypesCollectionPubSubTypes.hpp>
+    #include <ddsrecorder_participants/common/types/dynamic_types/v2/DynamicTypesCollection.hpp>
+    #include <ddsrecorder_participants/common/types/dynamic_types/v2/DynamicTypesCollectionPubSubTypes.hpp>
 #endif // if FASTRTPS_VERSION_MAJOR <= 2 && FASTRTPS_VERSION_MINOR < 13
 
 #include <ddsrecorder_participants/constants.hpp>
