+++ ddsreplayer/src/cpp/tool/DdsReplayer.cpp
@@ -34,11 +34,11 @@
 #include <ddspipe_core/types/dynamic_types/types.hpp>
 
 #if FASTRTPS_VERSION_MAJOR <= 2 && FASTRTPS_VERSION_MINOR < 13
-    #include <ddsrecorder_participants/common/types/v1/DynamicTypesCollection.hpp>
-    #include <ddsrecorder_participants/common/types/v1/DynamicTypesCollectionPubSubTypes.hpp>
+    #include <ddsrecorder_participants/common/types/dynamic_types/v1/DynamicTypesCollection.hpp>
+    #include <ddsrecorder_participants/common/types/dynamic_types/v1/DynamicTypesCollectionPubSubTypes.hpp>
 #else
-    #include <ddsrecorder_participants/common/types/v2/DynamicTypesCollection.hpp>
-    #include <ddsrecorder_participants/common/types/v2/DynamicTypesCollectionPubSubTypes.hpp>
+    #include <ddsrecorder_participants/common/types/dynamic_types/v2/DynamicTypesCollection.hpp>
+    #include <ddsrecorder_participants/common/types/dynamic_types/v2/DynamicTypesCollectionPubSubTypes.hpp>
 #endif // if FASTRTPS_VERSION_MAJOR <= 2 && FASTRTPS_VERSION_MINOR < 13
 
 #include <ddsrecorder_participants/constants.hpp>
